test: remove unnecessary dtype specification

+++ tests/constellations/test__constellations.py
@@ -55,7 +55,7 @@ def test_constellation_equivalence_properties(const: komm.abc.Constellation):
 def test_constellation_equivalence_mod_demod(const: komm.abc.Constellation, snr):
     ref = komm.Constellation(const.matrix)
     channel = komm.AWGNChannel(signal_power=float(const.mean_energy()), snr=snr)
-    indices = np.random.randint(0, const.order, size=100, dtype=int)
+    indices = np.random.randint(0, const.order, size=100)
     received = channel.transmit(const.indices_to_symbols(indices))
     np.testing.assert_allclose(
         const.indices_to_symbols(indices),
@@ -74,7 +74,7 @@ def test_constellation_equivalence_mod_demod(const: komm.abc.Constellation, snr)
 
 def test_constellation_equivalence_high_snr(const: komm.abc.Constellation):
     ref = komm.Constellation(const.matrix)
-    indices = np.random.randint(0, const.order, size=100, dtype=int)
+    indices = np.random.randint(0, const.order, size=100)
     symbols = const.indices_to_symbols(indices)
     np.testing.assert_equal(const.closest_indices(symbols), indices)
     np.testing.assert_equal(ref.closest_indices(symbols), indices)
